moved the functions back to their files
